+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ## 1.8.1 - Unreleased
 
+* Fixed [#425](https://github.com/NeoVintageous/NeoVintageous/issues/425): Reloading vintageousrc should unload existing mappings
 * Fixed [#422](https://github.com/NeoVintageous/NeoVintageous/issues/422): `{Visual}y` should highlight the selection (HighlightedYank)
 
 ## 1.8.0 - 2019-01-23